update to 1.16-pre2

+++ src/main/java/user11681/limitless/mixin/ItemStackMixin.java
@@ -11,6 +11,6 @@ import org.spongepowered.asm.mixin.injection.Redirect;
 public abstract class ItemStackMixin {
     @Redirect(method = "addEnchantment", at = @At(value = "INVOKE", target = "Lnet/minecraft/nbt/CompoundTag;putShort(Ljava/lang/String;S)V"))
     protected void removeLimit(final CompoundTag compoundTag, final String key, final short value, final Enchantment enchantment, final int level) {
-        compoundTag.putInt("lvl", level);
+        compoundTag.putShort("lvl", (short) level);
     }
 }
